refactor: Split index.js into App.js
